+++ camping_server/bot/slackbot.py
